refactor: project selection dialog
